correct print varaidic args

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -16,7 +16,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 	va_start(list, n);
 	while (x <= n)
 	{
-		printf("%d", va_arg(list, int));
+		printf("%d", va_arg(list, long));
 		x++;
 		if (separator != NULL)
 			printf("%s", separator);
